Fix loading the gem.
wikidatum/client is the first require, and so Ruby gets confused (I
guess?) if the Wikidatum module isn't defined and we try to
define `class Wikidatum::Client`.

So this fixes that by explicitly defining the module in client.rb.

Also remove the Gemfile and Rakefile from the files included in the
built gem.